refactor: decompose useAccountAuth into repository/routes/session modules
hooks/useAccountAuth.ts was 786 lines. Split it into:
- utils/accountAuthRoutes.ts: route/URL builders (normalizeAccountDisplayName,
  normalizeAccountFlowReturnTo, buildAccountAuthRoute, buildAccountAuthRedirectUrl)
- utils/accountRepository.ts: Supabase persistence (bootstrapAccountRow,
  saveAccountDisplayName, load/saveAccountSyncedSettings)
- hooks/useAccountSessionSync.ts: session restore + auth-state-change sync
- hooks/useAccountSettingsSync.ts: the store-subscription settings-sync effect

useAccountAuth.ts keeps AccountAuthProvider + auth actions and re-exports
every symbol it previously exported (test/consumer import paths unchanged).
Pure move, no logic edits. All files now under 400 lines (largest: 356).
